+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 * 优化代码结构，新增单isolate时可以不使用队列保存task直接发送到isolate
+
+## 1.0.2
+
+* 增加isolate初始化时存入map values
+* keepAliveTime为0时优化isolate的退出机制